Add calendar member variable to StudentManager
The StudewntManager#createPerson() API requires the current year when
creating the ID. Current implementation creates an instance of
GregorianCalendar and gets the year from that instance. This makes
testing more difficult.

This change uses dependency injection to create an instance variable
calendar and get the year from the externally-supplied instance.